App is now more convenient for end user to quickly generate file
By having the app frame and limit the questions, the user doesn't have to bother with typing in a lot of the information themselves. If the user really wants to change things granularly, they can do it after the file is created -- this app is for quick, easy professional templates, not for wide customization.